add pip packaging
- uses flit as a build backend
- moves the python/ directory to scalu/ to better conform to standards
- changes import paths to match new packaging format
