--- python/src/frontend/parser/parser.py
@@ -1,6 +1,0 @@
-
-import src.frontend.parser.expect as expect
-
-def parse(tokens):
-    return expect.parse(tokens)
-
