+++ scalu/src/frontend/parser/parser.py
@@ -1,0 +1,6 @@
+
+import scalu.src.frontend.parser.expect as expect
+
+def parse(tokens):
+    return expect.parse(tokens)
+
