+++ scalu/src/frontend/utility/utility.py
